Wrap inputs in simple components
This is probably overkill, and it's probably also overkill to have a common reused input box type, but it's nice to have both number and text inputs consume a shared input type, so they can just decorate it with behavior specific to handling their data type. Also, I'm using this as an exercise to practice composition over inheritance. Whenever I have something that reaches the level of complexity that I want to implement an abstraction, I should make an effort to always look for composition before inheritance. The one annoying part about composition is that you have to pass everything through to the abstraction that gets consumed.